requirements.txt
streamlit>=1.29.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.41.0
huggingface-hub>=0.22.0
sentencepiece>=0.1.99
timm>=0.9.12
accelerate>=0.30.0
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
